Migrate some utils to module utils
