+++ plugins/modules/zos_copy.py
@@ -67,9 +67,6 @@ options:
       - When set to C(true), the module creates a backup file or data set.
       - The backup file name will be returned on either success or failure of
         module execution such that data can be retrieved.
-      - Note that a PDS/PDSE member can not be backed up to the same PDS/PDSE.
-        If destination is a partitioned data set member, the entire partitioned
-        data set will backed up.
     type: bool
     default: false
     required: false
@@ -85,8 +82,10 @@ options:
         file will be the destination file or path name appended with a
         timestamp, e.g. C(/path/file_name.2020-04-23-08-32-29-bak.tar).
       - If the destination is an MVS data set, it will be a data set with a random
-        name generated by calling the ZOAU API. The MVS backup data set recovery 
+        name generated by calling the ZOAU API. The MVS backup data set recovery
         can be done by renaming it.
+      - If C(dest) is a data set member and C(backup_file) is not provided, the data set
+        member will be backed up to the same partitioned data set with the name "MEMBAK".
     required: false
     type: str
   force:
@@ -1124,41 +1123,6 @@ def get_file_checksum(src):
     return hash_digest.hexdigest()
 
 
-def extract_dsname(data_set):
-    """Extract the actual name of the data set from a given input source
-
-    Arguments:
-        data_set {str} -- Input data set name
-
-    Returns:
-        {str} -- The actual name of the data set
-    """
-    result = ""
-    for c in data_set:
-        if c == '(':
-            break
-        result += c
-    return result
-
-
-def extract_member_name(data_set):
-    """Extract the member name from a given input source
-
-    Arguments:
-        data_set {str} -- Input source name
-
-    Returns:
-        {str} -- The member name
-    """
-    start = data_set.find('(')
-    member = ""
-    for i in range(start + 1, len(data_set)):
-        if data_set[i] == ')':
-            break
-        member += data_set[i]
-    return member
-
-
 def cleanup(src_list):
     """Remove all files or directories listed in src_list. Also perform
     additional cleanup of the /tmp directory.
@@ -1229,9 +1193,9 @@ def run_module(module, arg_def):
     # in the form DATA.SET.NAME(MEMBER). When this is the case, extract the
     # actual name of the data set.
     # ********************************************************************
-    dest_name = extract_dsname(dest)
-    src_name = extract_dsname(src) if src else None
-    member_name = extract_member_name(src) if src_member else None
+    dest_name = data_set.extract_dsname(dest)
+    src_name = data_set.extract_dsname(src) if src else None
+    member_name = data_set.extract_member_name(src) if src_member else None
 
     conv_path = None
     src_ds_vol = None
@@ -1304,7 +1268,7 @@ def run_module(module, arg_def):
             module.exit_json(note="Destination exists. No data was copied")
 
         if backup or backup_file:
-            backup_file = backup_data(dest_name, dest_ds_type, backup_file)
+            backup_file = backup_data(dest, dest_ds_type, backup_file)
     # ********************************************************************
     # If destination does not exist, it must be created. To determine
     # what type of data set destination must be, a couple of simple checks
